move 'calculate end time' helper to globals + add description
